requirements.txt
alembic>=1.10.0
python-dotenv>=1.0.0
pytest>=7.0.0
SQLAlchemy>=1.4.0,<2.0.0
pymysql>=1.0.0
cryptography>=40.0.0
PyJWT>=2.8.0
Flask>=2.3.0
bleach>=6.0.0
Flask-Cors>=4.0.0
